metric comp: removed unused utils functions and moved the useful utils function into a new specific module
