+++ dadapy/utils_/utils.py
@@ -282,111 +282,6 @@ def _compute_binomial_cramerrao(d, k, r, n):
 
 
 # --------------------------------------------------------------------------------------
-# Functions used in the metric_comparisons module
-# --------------------------------------------------------------------------------------
-
-
-def _return_ranks(dist_indices_1, dist_indices_2, k=1):
-    """Finds all the ranks according to distance 2 of the kth neighbours according to distance 1.
-
-    Args:
-        dist_indices_1 (np.ndarray(int)): N x maxk matrix, nearest neighbours according to distance1
-        dist_indices_2 (np.ndarray(int))): N x maxk_2 matrix, nearest neighbours according to distance2
-        k (int): order of nearest neighbour considered for the calculation of the conditional ranks, default is 1
-
-    Returns:
-        np.array(int): ranks according to distance 2 of the first neighbour in distance 1
-
-    """
-    assert dist_indices_1.shape[0] == dist_indices_2.shape[0]
-
-    N = dist_indices_1.shape[0]
-    maxk_2 = dist_indices_2.shape[1]
-
-    conditional_ranks = np.zeros(N)
-
-    for i in range(N):
-        idx_k_d1 = dist_indices_1[i, k]
-
-        wr = np.where(idx_k_d1 == dist_indices_2[i])
-
-        if len(wr[0]) == 0:
-            conditional_ranks[i] = np.random.randint(maxk_2, N)
-        else:
-            conditional_ranks[i] = wr[0][0]
-
-    return conditional_ranks
-
-
-def _return_imbalance(dist_indices_1, dist_indices_2, k=1):
-    """Compute the information imbalance between two precomputed distance measures.
-
-    Args:
-        dist_indices_1 (np.ndarray(int): nearest neighbours according to distance1
-        dist_indices_2 (np.ndarray(int): nearest neighbours according to distance2
-        k (int): order of nearest neighbour considered for the calculation of the imbalance, default = 1
-
-    Returns:
-        (float): information imbalance from distance 1 to distance 2
-
-    """
-    assert dist_indices_1.shape[0] == dist_indices_2.shape[0]
-
-    N = dist_indices_1.shape[0]
-
-    ranks = _return_ranks(dist_indices_1, dist_indices_2, k=k)
-
-    imb = np.mean(ranks) / (N / 2.0)
-
-    return imb
-
-
-def _return_imb_between_two(X, Xp, maxk, k, ltype="mean"):
-    nbrsX = NearestNeighbors(
-        n_neighbors=maxk, algorithm="auto", metric="minkowski", p=2, n_jobs=1
-    ).fit(X)
-
-    nbrsXp = NearestNeighbors(
-        n_neighbors=maxk, algorithm="auto", metric="minkowski", p=2, n_jobs=1
-    ).fit(Xp)
-
-    _, dist_indices_Xp = nbrsXp.kneighbors(Xp)
-
-    _, dist_indices_X = nbrsX.kneighbors(X)
-
-    nXp_X = _return_imbalance(dist_indices_Xp, dist_indices_X, k=k, dtype=ltype)
-    nX_Xp = _return_imbalance(dist_indices_X, dist_indices_Xp, k=k, dtype=ltype)
-
-    return nX_Xp, nXp_X
-
-
-def _return_imb_linear_comb_two_dists(
-    dist_indices_Y, d1, d2, a1, maxk, k=1, ltype="mean"
-):
-    dX = np.sqrt((a1**2 * d1**2 + d2**2))
-
-    dist_indices_X = np.asarray(np.argsort(dX, axis=1)[:, 0 : maxk + 1])
-
-    nX_Y = _return_imbalance(dist_indices_X, dist_indices_Y, k=k, dtype=ltype)
-    nY_X = _return_imbalance(dist_indices_Y, dist_indices_X, k=k, dtype=ltype)
-
-    return nX_Y, nY_X
-
-
-def _return_imb_linear_comb_three_dists(
-    dist_indices_Y, d1, d2, d3, a1, a2, maxk, k=1, ltype="mean"
-):
-    dX = np.sqrt((a1**2 * d1**2 + a2**2 * d2**2 + d3**2))
-
-    dist_indices_X = np.asarray(np.argsort(dX, axis=1)[:, 0 : maxk + 1])
-
-    nX_Y = _return_imbalance(dist_indices_X, dist_indices_Y, k=k, dtype=ltype)
-    nY_X = _return_imbalance(dist_indices_Y, dist_indices_X, k=k, dtype=ltype)
-
-    return nX_Y, nY_X
-
-
-# --------------------------------------------------------------------------------------
 # Others
 # --------------------------------------------------------------------------------------
